+++ src/queue.cpp
@@ -10,7 +10,7 @@ Queue* queue_new()
     return new_queue;
 }
 
-int queue_is_empty(Queue* q)
+long queue_is_empty(Queue* q)
 {
     if (q->length == 0)
         return TRUE;
@@ -34,7 +34,7 @@ void queue_push_tail(Queue* q, void* item)
     q->length++;
 }
 
-int queue_get_length(Queue* q) { return q->length; }
+long queue_get_length(Queue* q) { return q->length; }
 
 void* queue_pop_head(Queue* q)
 {
