+++ src/queue.h
@@ -6,12 +6,12 @@ typedef struct Queue_Item {
 typedef struct Queue {
     Queue_Item* first_item;
     Queue_Item* last_item;
-    int length;
+    long length;
 } Queue;
 
 Queue* queue_new();
-int queue_is_empty(Queue* q);
+long queue_is_empty(Queue* q);
 void queue_push_tail(Queue* q, void* item);
-int queue_get_length(Queue* q);
+long queue_get_length(Queue* q);
 void* queue_pop_head(Queue* q);
 void queue_free(Queue* q);
